types: clean up uintX type insanity
Crazy types all over the place. Lines and columns are always referenced
by an int now.

+++ ansi.go
@@ -67,7 +67,7 @@ func (t *Terminal) handleScreenState(readChan chan measuredRune) bool {
 
 		// hide cursor?
 		buffer := t.getActiveBuffer()
-		buffer.resetVerticalMargins(uint(buffer.viewHeight))
+		buffer.resetVerticalMargins(buffer.viewHeight)
 		buffer.SetScrollOffset(0)
 
 		// Fill the whole screen with E's
